Move version selection system to its own dialog.
